example 항목별로 분류
