+++ example/food/myUtils.js
@@ -1,0 +1,7 @@
+const util = require('util');
+
+const myLogger = (res) => {
+  console.log(util.inspect(res, {showHidden: false, depth: null}));
+};
+
+exports.myLogger = myLogger;
