+++ lib/common/baseUrl.js
@@ -1,6 +1,6 @@
 module.exports = {
   NOTICE_LIST: 'http://m.ssu.ac.kr/html/themes/m/html/notice_univ_list.jsp?curPage=&sCategory=',
   SOONGGURI_JSON: 'http://soongguri.com/menu/m_menujson.php?fkey={DAY_OF_WEEK}',
-  SOONGGURI_DESKTOP_URL: '://www.soongguri.com/main.php?mkey=2&w=3&l=3',
+  SOONGGURI_DESKTOP_URL: 'http://www.soongguri.com/main.php?mkey=2&w=3&l=3',
   SSU_DORM_URL: 'http://ssudorm.ssu.ac.kr/SShostel/mall_main.php?viewform=B0001_foodboard_list&gyear={YEAR}&gmonth={MONTH}&gday={DAY}',
 };
