+++ lib/food/errors.js
@@ -33,10 +33,17 @@ class UnexpectedParameterDate extends BaseError {
   }
 }
 
+class NoFoodToday extends BaseError {
+  constructor(message) {
+    super(`No data ${message} on school server checkout ${urls.SOONGGURI_DESKTOP_URL}`);
+  }
+}
+
 module.exports = {
   SsuDormConnectionError,
   SoongguriJsonConnectionError,
   SoongguriDesktopConnectionError,
   UnexpectedParameterDayOfWeek,
   UnexpectedParameterDate,
+  NoFoodToday,
 };
