+++ lib/food/parsers.js
@@ -6,6 +6,7 @@ const {
   replaceContinuousSpaceToNewLine,
   removeBetweenParentheses,
 } = require('./cafeteriaUtils');
+const errors = require('./errors');
 
 function parseOfficer(body) {
   // 교식
@@ -16,6 +17,9 @@ function parseOfficer(body) {
   //   '석식1': [ '참치마요덮밥', '볶음김치계란국요구르트' ] }
   const json = JSON.parse(body);
   const facultyFoods = json['교직원식당'];
+  if (facultyFoods === undefined) {
+    throw new errors.NoFoodToday('교직원식당');
+  }
   const ret = [];
   Object.keys(facultyFoods).forEach((time) => {
     const $ = cheerio.load(facultyFoods[time]);
@@ -59,6 +63,9 @@ function parseStudent(body) {
    '흰밥' ] } */
   const json = JSON.parse(body);
   const pupilFoods = json['학생식당'];
+  if (pupilFoods === undefined) {
+    throw new errors.NoFoodToday('학생식당');
+  }
   const ret = [];
   Object.keys(pupilFoods).forEach((time) => {
     const $ = cheerio.load(pupilFoods[time]);
@@ -109,6 +116,9 @@ function parseResidence(response) {
 function parseFoodCourt(body) {
   const json = JSON.parse(body);
   const foodCourtFoods = json['푸드코트'];
+  if (foodCourtFoods === undefined) {
+    throw new errors.NoFoodToday('푸드코트');
+  }
   const priceReg = /\d\.\d$/g;
   const ret = [];
   Object.keys(foodCourtFoods).forEach((section) => {
@@ -158,6 +168,9 @@ function parseFoodCourt(body) {
 function parseSnackCorner(body) {
   const json = JSON.parse(body);
   const foodCourtFoods = json['스넥코너'];
+  if (foodCourtFoods === undefined) {
+    throw new errors.NoFoodToday('스넥코너');
+  }
   const ret = [];
   const priceReg = /\d\.\d$/g;
   Object.keys(foodCourtFoods).forEach((section) => {
@@ -215,7 +228,8 @@ function parsePrice(response, menuList) {
       $('span').each((i, elem) => {
         if ($(elem).text().length < 50) {
           if ($(elem).text().includes(menu)) {
-            const priceIncludedString = $(elem).parent().parent().parent().text();
+            const priceIncludedString = $(elem).parent().parent().parent()
+              .text();
             const price = priceIncludedString.match(priceReg);
             if (price === null) return;
             prices.push(price);
